requirements.txt
Flask==2.0.1
face-recognition==1.3.0
Pillow==8.3.2
numpy==1.21.2
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
dlib==19.22.0